+++ src/Caster.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Spatie\DataTransferObject;
+
+interface Caster
+{
+    public function cast(mixed $value): mixed;
+}