+++ src/Exceptions/ValidationException.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace Spatie\DataTransferObject\Exceptions;
+
+use Exception;
+use Spatie\DataTransferObject\DataTransferObject;
+
+class ValidationException extends Exception
+{
+    public function __construct(DataTransferObject $dataTransferObject, array $validationErrors)
+    {
+        $className = $dataTransferObject::class;
+
+        $messages = [];
+
+        foreach ($validationErrors as $fieldName => $errorsForField) {
+            /** @var \Spatie\DataTransferObject\Validation\ValidationResult $errorForField */
+            foreach ($errorsForField as $errorForField) {
+                $messages[] = "\t - `{$className}->{$fieldName}`: {$errorForField->message}";
+            }
+        }
+
+        parent::__construct("Validation errors:" . PHP_EOL . implode(PHP_EOL, $messages));
+    }
+}