+++ src/FlexibleDataTransferObject.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Spatie\DataTransferObject;
-
-abstract class FlexibleDataTransferObject extends DataTransferObject
-{
-    protected bool $ignoreMissing = true;
-}