+++ tests/Dummy/ComplexObject.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Spatie\DataTransferObject\Tests\Dummy;
+
+class ComplexObject
+{
+    public function __construct(
+        public string $name,
+    ) {
+    }
+}