+++ tests/Dummy/ComplexObjectWithCaster.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Spatie\DataTransferObject\Tests\Dummy;
+
+use Spatie\DataTransferObject\Attributes\CastWith;
+
+#[CastWith(ComplexObjectWithCasterCaster::class)]
+class ComplexObjectWithCaster
+{
+    public function __construct(
+        public string $name,
+    ) {
+    }
+}